dialects: (aie) Add aie dialect (#1763)
Port of the AMD Xilinx AIE dialect for programming the AIEs on the AMD
Xilinx Versal FPGA architecture. AIE is a hardened systolic array
present in the Versal devices. The dialect describes netlists of AIE
components and it can be lowered to the processor's assembly using the
vendor's compiler. A description of the original dialect can be found
here https://xilinx.github.io/mlir-aie/AIE.html