testing: (snitch_stream) modify test to minimise upcoming PR diff (#2142)
PR 3/4 of replacing the snitch_stream.stride_pattern op with an
attribute

The filecheck test most impacted by this refactoring is the lowering,
this PR moves some code around to make the upcoming diff as small as
possible. It conceptually adds a scenario to test, configuring each dm
individually when the patterns aren't the same, and configuring them all
using dm=31 when they are the same.