dialects: (snitch_stream) replace stride pattern op with attribute (#2143)
PR 4/4 of replacing the snitch_stream.stride_pattern op with an
attribute

The operation is removed, and the stride pattern is specified as an
attribute on the streaming region op. When lowering the streaming
region, the dm is inferred.

Note that the assembly that is returned by our lowering is entirely
unchanged after this refactor.