backend: (riscv) correctly handle f64 moves when lowering (#2156)
We currently incorrectly insert fmv.s when moving float registers, even
when the actual value stored is f64. This is usually at the boundary
between the standard and riscv dimensions, where the combination of
casting to a register type and moving to/from an "a" register loses the
relevant type information. This PR amends some helpers in the riscv
backend to take into account the floating-point type of the operand when
moving floating point values.