dialects: (memref_stream) add memref_stream.generic (#2148)
Note stacked PR

This PR introduces the memref_stream equivalent of the linalg generic
operation. The main differences are that this one has an extra property
`bounds`, which contains the bounds of the loop nest, and that the
operands may be streams, in which case they'll be lowered to stream
reads and writes instead of loads and stores.