core: Allow to get the expected base of an IRDL constraint (#2175)
This PR adds `get_unique_base` on `AttrConstraint` that returns the
unique base (if it exists) that can
satisfy the constraint. In other terms, it will return the only
`Attribute` type that can be accepted by the
IRDL attribute constraint.

The main application of this will be in the declarative format. In MLIR,
if a type is already known to be
of a specific base, then it will not print its own type in the custom
format.