core: add Hasher and HashableModule (#2210)
These utils are planned to be used in `xdsl-gui`, but may well be useful
outside of that project. The Hasher contains logic I've used a lot
before, it seems to work well enough. In the future, we might want a
more efficient way to compute equality/inequality for modules, but we
might as well add a working helper class now and optimise later.