dialects: Fix inline_asm unitattrs to not be properties (#2223)
@superlopuh pointed out that #2203 contains an erroneous implementation
of a UnitAttr being implemented as properties. This is incorrect, since
upstream prints the UnitAttrs as attributes and properties.
It was a bit confusing because some other ops in the LLVM dialect have
the same mistake right now.
This happens under the radar because in testing, properties are printed
as attributes.
This PR fixes the inline_asm operation.