interactive: make some structural changes that will help tree PR 2 (#2217)
To reuse the functionality we already have in the tool where we ask the
user to input pass arguments, we make some structural changes so that we
can reuse it in future with the tree structure (node expansion).

self.get_pass_arguments() function computes the pass tuple with the
arguments, in a format ready to be added to the pass_pipeline. Actually
adding it to the pass_pipeline happens in the update_pass_pipeline()
function instead of in the self.get_pass_arguments() function (the way
it used to be).

Since functions to do with screens cannot return the variable (i.e the
add_pass_with_arguments_to_pass_pipeline), I create a global
`current_argument_pass` variable that holds the variable so that it can
be accessed later.