interactive: add update_selected_passes_list_view helper function PR 3 (#2218)
Now starts a series of mini-pr's that will add helper functions that
will help make the multi-level tree traversal. please trust the vision
that we need these helper functions!

update_selected_passes_list_view() is passes a pipeline_variable because
in the future we will not want it to contain the last pass in the
pipeline (as that will be the root of the tree).