interactive: make root of tree to change from input to last selected pass  PR 4 (#2219)
PR 2

`expand_node` takes a root node of the tree of passes, as well as a
children_nodes_list and added one level to the tree (i.e. adds the
children_nodes_list as the children of the root node).

`update_root_of_passes_tree` helper function helps keep track of what
the root of a tree is when the pass_pipeline is updated (i.e. a pass is
selected and root of the tree changes).

While these functions are already useful now, they will also be more
useful in future with tree building!