interactive: remove passes that raise errors from condensed list (#2207)
I have a feeling this should have been the case from the get-go. The
main reason to condense is to show the available transformations, to
make it quick and easy to go towards your target. There are three
reasons why a pass might fail:
1. it's mlir-opt and it has weird stdout/stderr behaviour that breaks
textual
2. it's a pass that requires parameters to run
3. it actually fails

I would say that none of these are useful by themselves to show in the
"quick and easy" list. If the user knows to use a certain pass, they can
always "uncondense" and pick the pass that they want, possibly with the
parameters they're interested int. The passes that currently require
parameters should probably either provide default parameters or provide
a range of options to try in the future, but that can be done by another
PR.