interpreter: add index_bitwidth to interpreter (#2230)
Some of the high-level function semantics depend on the width of the
pointer type. This PR adds this information to the interpreter, making
it accessible for interpreter impls. Currently none of the impls
leverage this, so we're no worse off than before with this PR in, and
can get to fixing the existing bugs.