testing: rename test-lower-linalg-to-snitch to test-lower-snitch-stream-to-asm (#2237)
The main idea is that we want to keep a part of this stable, since we
still have some tests defined as snitch_stream in the RISC-V backend
paper experiments. I'd like to propose having two passes, one for linalg
to snitch stream, and a second for lowering snitch stream down to
assembly.