transformations: simplify stride patterns when lowering memref_stream (#2253)
Note stacked PR

When lowering convolutions, we run out of stream dimensions, hence this
PR, which lowers the number of dimensions in the stream. In a future PR,
we'll only convert a memref operand to a `generic` to a stream when we
know that the stride pattern will be less than some parameter. For now,
we just convert first and then hope for the best.