dialects: (riscv) add index to registers (#2321)
The identity of registers is currently defined by a string. This can be
problematic as RISC-V registers have alternative spellings which are
then considered as unequal.

This PR is a first step towards fixing that issue by adding an optional
`index` parameter to the type. The idea is to have a unique index that
identifies the register in the future if it is a predefined register.
Non-predefined registers such as the infinite `j[0-9]+` registers set
`index` to `None` to signal that the spelling should be used for
equality.

A future PR will add the `__eq__` implementation after the
representation has proven itself.

Part of https://github.com/xdslproject/xdsl/issues/1760