core: Make print_ssa_value return chosen name (#2370)
Sometimes it is useful to know when printing what name has been chosen
to print an SSA value (for example when SSA value names should map to
names in an interface like in CIRCT's hardware modules). This PR
introduces this feature to the printer. This also adjusts the type
bounds on `print_list` so it remains happy with having `print_ssa_value`
as its argument.