transformations: (reconcile-unrealized-casts) Disable warnings in pattern by default (#2378)
The pattern for unrealized casts reconciliation emits a lot of warnings
when it fails to match. While this makes sense as a pass application, it
clutters output in patterns where reconciliation is expected to fail.
This PR makes the emission of warning togglable, disabled by default in
the pattern and enabled in the pass.