dialects: (hw) Add `hw.module` and `hw.output` operations (#2366)
This PR introduces the `hw.module` and its companion `hw.output` from
CIRCT, while adding various small enhancements to the general
infrastructure to accommodate them. ModuleOps are used to specify
self-contained hardware modules, akin to functions in software.

A `hw.module` looks like the following:
```
hw.module @name<param1: i1, param2: i2>(in %a: i8, in %b: i9, out c: i8) {
  %res = arith.constant 0 : i8
  hw.output %res : i8
}
```
While hardware modules have similar semantics to functions, both inputs
and outputs in `hw.module` are defined within the parenthesis, as
*ports*. This is because some ports may be both input and output in
CIRCT (currently not supported). Modules can optionally receive static
parameters (in the example, `param1` and `param2`). Those are akin to
C++ value templates or Rust const generics, and are useful to factor
similar modules together.

`DirectionAttr`, `ModulePort` and `ModuleType` are defined to represent
the type of a module and its ports. `ParamDeclAttr` is an attribute to
represent the parameter definitions on the module.

The `hw.output` operation is the terminator for `hw.module`, setting the
value of the output ports.