dialects: (x86) PR5_1 - single operand instructions - one source (#2398)
Added support for single operand instructions and the following specific
cases: idiv, pop, push and not. I added some tests for the verification
of the single operand instructions given that the one operand can be a
destination, a source or both.

idiv is a bit of a special case since it technically has 2 destination
registers: RAX for quotient and RDX for remained, but they're not
specified by the programmer but rather built-in. I'd be happy to discuss
what the best approach to this is and if it should create SSA values.