dialects: (x86) - Reformats and changes in naming convention (#2487)
I've done some changes to the naming conventions following the
suggestions of @tarinduj. This includes removing the
single/doubleOperand parent classes (issue #2466), writing the class
names in the format [Result]-[Operands]-Operation, and changing the
individual names of the operations to prepare for the different
configurations those instructions can exist in. The prefix signifies the
type of the operand passed into the assembly instruction and is removed
during assembly emission.