dialects: (hw) Revert HWModuleOp's usage of InnerRefNamespaceTrait (#2498)
While upstream MLIR decouples SymbolTables from operations, this is not
the case in xDSL. As such, whenever one wants to reuse symbol table
functionality, they have to declare an operation as a SymbolTable op.
This was done on HWModuleOp to reuse SymbolTable functionality for the
InnerSymbolTable, but in contrast to how it is done in upstream CIRCT,
this thus requires the InnerSymbolTable to be a regular SymbolTable.
This in turn blocks regular symbol referencing outside of the
InnerSymbolTable.

This PR reverts this change for now. This should be reimplemented once
xDSL supports decoupled symbol tables (or InnerRefNamespaceTrait is
updated accordingly).