core: Factor out visibility keyword parsing (#2507)
This PR factors out the parsing of visibility keyword into core (like in
MLIR) as visibility keywords are a property of the symbol
infrastructure. While the symbol lookup in xDSL does not take it into
account yet, it is useful to have it be part of the symbol
infrastructure itself as those keywords are used in many places (for
example, in func.func or hw.module).