testing: update matmul bottom-up test (#2547)
This is almost exactly the same kernel as we currently have as the
target in the snitch paper. It achieves perfect utilisation in the
steady state. This is achieved by interleaving four dot products in our
matmul, which avoids the stalls that occur when an instruction depends
on the result of the previous instruction.

Note stacked PR