core: Block inlining API improvements (#2572)
The block argument replacement was added to the core function, but only
to one of the API ones. This PR adds it to all of them.

Also minor tweaks:
- Block inlining currently asserts the inlined block is part of a
region, only to detach it. Switch to simply detaching it if attached.
- Block inlining currently asserts that the right number of block
argument replacement values is passed. Switch to just not do argument
replacement if no such values are passed.