interpreter: use function type when parsing xdsl-run arguments (#2536)
There was an aspect of the interpreter that I didn't account for when
opening the original PR that allowed passing function arguments, which
is that different levels of abstractions use different runtime
representations of the same thing. Notably, buffers are represented by
`ShapedArray` for values of type `MemrefType`, and with `RawPtr` for
RISC-V register values. This PR adds an extra parameter to the
`impl_attr`-annotated impls, and changes the interpreter logic to use
those to find the implementation, instead of the type of the parameter
passed via the command line.

Note stacked PR