transformations: Implement generic CSE. (#2631)
Implement generic CSE mimicking MLIR's.
We do not have the notion of Region kind, so this implementation is
pessimisticly handling everything as a Graph region. That means it is
slightly less powerful than MLIR's, but as safe.

The included file check is exactly the one in MLIR, with the same
expected output, unless specified otherwise on top of differing
examples.

The only unsupported cases are regarding side effects; given that we
only have `Pure` in xDSL (i.e., no operation of this type has any side
effect ever), this does not implement more subtle cases.
Those are CSEing consecutive read-only side-effects, if not interrupted
by write side-effects, and probably more importantly: recursive side
effect. (A very scary way to say "I have side effect if and only if my
regions contain side effects :shrug: ). That would allow CSEing
side-effect-less conditionals and what not.