backend: (riscv) Don't make unreserved registers available (#2620)
This is PR 2/2, note stacking.

This was a tricky one to wrap my head around. There is a feature of the
register allocator that we added in November, which prevents registers
used as iteration arguments from being reused inside the loop, by
"reserving them", meaning incrementing a counter. Currently, when the
counter reaches 0, the register is marked as available for reuse. The
issue is that the register may have been marked as unavailable while in
use, so unreserving it may result in incorrect behaviour. This PR
changes this, and adds an assert to check that a register wasn't reused
despite being reserved.