transforms: Add function-constant-pinning pass to specialize functions (#2389)
This adds the `function-constant-pinning` pass that takes annotated
functions and creates specialized versions of them that pin a value to a
constant.

Currently only supports pinning at the top level of a function. Also
other limitations, that will be documented on a train ride soon!

See the filecheck for examples on how the pass works!