misc: Implement and test dynamic IRDL dialect registering. (#2654)
This is an implementation - through our interpreter infrastructure - of
IRDL dialect registering. That is, at runtime, read an IRDL dialect in
MLIR form, and register the described dialect to use right away.

NB: This only implements the steps described here; avoiding any more
involved discussion abut how exactly to integrate this in `xdsl-opt` or
hack in Python import mechanism and what not. To be continued!

---------

Co-authored-by: Sasha Lopoukhine <superlopuh@gmail.com>
Co-authored-by: Mathieu Fehr <mathieu.fehr@gmail.com>