core: implement RangeConstraint and use it without functional change (#2675)
Had this in mind for a while now. This PR offers an implementation of
RangeConstraint: a constraint not over single attributes, but over
*sequences* of attributes.

Make IRDL operands and results use RangeConstraint instead of
AttrConstraint internally. For a non-functional change, implement two
basic RangeConstraints:

- `RangeOf`: constrain an attribute sequence to any number of attributes
respecting the passed AttrConstraint.
- `SingleOf` : constrain an attribute sequence to a single attribute
respecting the passed AttrConstraint.

I.e., `RangeOf` simply reflects the current behaviour of
variadic/optional operands/results, and `SingleOf` redundantly mimic the
current behaviour of non-variadic operands/results. (IRDL already checks
lengths! The redundancy is only here for clarity, and paving the way to
different choices if we want. Not sure we'll keep it!)

So all the currently used AttrConstraint in dialects are implicitly
wrapped in either of those RangeConstraints, for an invisible first
introduction of the concept :slightly_smiling_face:

Why bother? We can now declaratively reason about much wider properties,
e.g, `scf.for` results types having to match element wise the iter_args
:slightly_smiling_face: Only needs the introduction of the corresponding
RangeVariables.

---------

Co-authored-by: Fehr Mathieu <mathieu.fehr@gmail.com>