core: unify (Pattern)Rewriter insertion and inlining APIs around InsertPoint (#2660)
This reduces the surface area of the Rewriter and PatternRewriter. Note
that it doesn't modify the `matched_op` helpers, partially so that the
diff doesn't baloon, and partially because I think they are actually
quite helpful as they are. At least I'd love to talk about the
matched_op helpers separately from this PR, if possible.

Resolves #859