backend: (csl) Slightly rework printer to promote certain values to inline expressions (#2728)
This patch makes the printer promote certain values to inlined
expressions (e.g. not creating a new `const x = 5;` variable for an
`arith.constant` operation, or for arithmetic expressions.

This is for now applied very carefully to a select few operations. Some
checks are performed, such as that the value is not referenced at any
point using the `AddressOf`operation, that the operation that produced
the value is pure, and more.