backend: (riscv) Add prologue epilogue insertion (#2752)
The RISC-V call ABI requires a few registers, if clobbered, to be
restored by the callee prior to the function returning. This has so far
not been implemented meaning that any C compiler calling an xDSL
generated function may cause undefined behaviour.

This PR adds a pass performing the required register saving and
restoring by pushing and popping them from the stack. The registers are
saved in the prologue at the beginning of the function and an epilogue
restoring them is emitted before every return operation.

The pass is required to run after register allocation to see the
clobbering of callee-preserved registers. It itself does not require
register allocation to be run afterward.