transformations: Support constant inits in memref_stream.generic lowering [2/3] (#2764)
This PR adds support for lowering `memref_stream.generic`s with inits to
loops. It also updates the bottom-up tests to leverage this feature. As
you can see, there is no functional change in the final assembly, modulo
a single register difference.

I updated the `insert_load` interface in the loop lowering helpers to
communicate the index of the operand that for which the load is being
inserted, this allows us to use the const value directly.

Note stacked PR.