transformations: Add `memref-stream-(infer|generalize)-fill` passes (#2776)
These let us detect fill operations, fold them if possible, and convert
them back to a form that we can lower if they could not be folded. Also,
updates the bottom_up tests to go all the way from linalg.generic to our
snitch assembly. Note that this is not the case yet for the matrix op,
as that one is interleaved and peeled, those are the two steps
remaining.

Note stacked PR.