documentation: allow marimo notebooks as documentation (#2787)
I like these a lot for local development, and have struggled for a while
with getting them to fail properly when the xDSL code updates. This PR
introduces both the a sample notebook, but also a new test script that
just runs all the notebooks in the marimo folder in headless mode,
failing on error. This should give us plenty of coverage and confidence
in the notebooks working properly.