core: Make print_string aware of indentation (#2809)
This PR makes the `print_string` function in the printer aware of
indentation, and will print newlines in strings accordingly.

For cases where this is not desired (such as printing a newline
character in the indentation-aware newline logic), `print_string_raw`
has been added to print the string ignoring indentation.

Additionally, a helper `indented` function for with statements has been
added to control indentation within a scope. Feedback on this would be
much appreciated, as a possible concern is that it may not be super
intuitive (it indents newlines, not lines).