dialects: (memref) add custom syntax for memref_subview (#2840)
I tried to match MLIR exactly in the syntax, and found a bug in the
stencil lowering, which lowers to an invalid memref subview. My
understanding is that the subview must be a subview, and so negative
offsets are not allowed.

I won't merge without someone from the stencil side of things approving
the PR, but also I'd love to use this for the linalg to snitch project,
so I'm hoping that we can resolve this soonish.