transformations: Add support for strided MemRefs in 'load' and 'store'
The lowering previously assumed an identity layout and errored if a strided layout was present.
Strided layouts are unfortunately very common as they are a direct result of tiling. This PR therefore adds support for them in both the store and load operations.

As a side-effect, the `memref.subview` implementation could also be heavily simplified.