transformations: handle strided layout attr in memref_stream lowering (#2866)
We currently treat all memrefs as raw pointers in our
memref_stream.streaming_region lowering, but that's not correct, we want
to take the strided layout into account when possible. This PR adds some
helpers to compute the appropriate stride map, and uses them to
construct the correct snitch stride pattern.