core: Add FixedbitWidthType Interface (#2904)
These functions to determine the bit/byte width of an element are useful
in many places. I think I have them implemented in a couple of different
places myself, so I think this abstraction is quite useful in this
regard. The functions were now defined in a pass, I simply moved them
over.