dialects: (snitch_stream) add repetition to stride pattern (#2927)
Without the repeat setting when the last offset is 0, Snitch just
fetches every time without caching the value, repeat avoids this.

CC @jorendumoulin @JosseVanDelm seems like more and more accelerator
config! One of the things that's somewhat annoying is that we now have
to defensively reset the repetition count, it would be great to avoid
this eventually.