dialects: (memref_stream) add doc and library call support (#2928)
We are using the `library_call` attribute to designate to which
accelerator we want to dispatch a certain linalg generic to. Right now,
this information gets lost when converting from linalg.generic to
memref_stream.generic, which is somewhat annoying. This PR adds support
for these generic attributes in the memref stream dialect.