backend: (snitch) unify all the test passes as a single end-to-end pass (#2955)
Now that we have a full end-to-end working pipeline that we are also
using in downstream projects, I think it makes sense to centralise it to
make it easier to play with. For example, the next PR will add an
optimisation level parameter, that will let us explore how the
performance changes with given optimisations disabled.