api: deprecate `print_string_raw` in favor of indent=0 (#2951)
This is an effort to minimise the API surface a little bit, and to make
it a little more efficient. When there are no newlines, it's wasteful to
create a list of them, so the first step when printing a string is now
to check whether they're there. Raw printing is equivalent to printing
with an indentation of 0, modulo next line callbacks. I don't think it's
likely to end up in a situation where the difference is visible, so
where the user prints a raw string and expects the next line callback to
not be executed, and instead be saved for the next line that is printed
with the non-raw printing, please let me know if I'm missing something.
In all other cases the behaviour is the same as before.