transform: (stencil-tensorize-z-dimension) Tensorize arith.constant directly (#2970)
Add support for directly tensorising `arith.constant`, e.g.
```
%0 = arith.constant dense<1.666600e-01> : tensor<510xf32>
```

The more generic way of tensorising any scalar operand (independent of
its source) would be `linalg.fill`ing the value a newly created empty
tensor. This still works as a generic way. However, in the special case
that the scalar operand is a constant, this is much shorter.

---------

Co-authored-by: n-io <n-io@users.noreply.github.com>