CI: use ruff to format xdsl (#3023)
Runs in `0.04s` on my computer.

The Ruff format is designed to be a drop-in replacement for black. As
you see, most of the diff is adding a new line at the start of every
class, which seems like a good change. I'll comment on the config files
with changes for easier reviewing.

Note stacked PR.