core: Add `BitEnumAttribute` (#2617)
The `BitEnumAttribute` is meant to be the equivalent of `BitEnumAttr` in
MLIR:
https://github.com/llvm/llvm-project/blob/e67f2cc3fc38cec2041cfb197ac4688ed3d16e7e/mlir/include/mlir/IR/EnumAttr.td#L266
It allows defining an attribute which in abstract is a set of flag
values where each flag value is defined within an enum.

The base class is meant to be easy to use like `EnumAttribute` and
automatically define suitable parsers and printers.
For an in-tree user that is converted by this PR see `FastMathAttr` in
the LLVM and arith dialects.

---------

Co-authored-by: Mathieu Fehr <mathieu.fehr@gmail.com>
Co-authored-by: Sasha Lopoukhine <superlopuh@gmail.com>
Co-authored-by: Joren Dumoulin <joren.dumoulin@kuleuven.be>