dialects: dmp: make strategies attributes and carry them in IR (#3050)
As of now, the decomposition strategies are dataclasses and the
information to construct them is expected to match in multiple bits of
the pipeline (actually fused in `distribute-stencil`, which happens to
do shape inference too, but I'm trying to decouple things here)

So instead, define them as attributes, and make the initial distribution
pass pop that attribute on the swaps. The swap shape inference then do
not need extra information to carry on.