transformations: (csl) use constants instead of params
`csl.param`s allow for flexibility we do not utilise whilst providing a
possibility of accidentally overriding a value which was not intended to
be overridden. In the `ModuleOp.get_param_value` function we even use
the value of the param as if it will always be there.

After this PR: `csl_wrapper.ParamAttribute` can no longer omit its value (now referred
to as "value" not "default")

In the layout module:
* All `param`s are now `arith.constant`s

In the program module:
* All params from the `csl_wrapper.ParamAttribute` are now
  `arith.constant`s
* Params from the layout module `yield` arguments are still
  `csl.param`s.