core: always print unregistered ops (#3303)
`scf.for` prints it's blocks without terminators if there are no
`iter_args`. Unfortunately the check for whether an operation is a
terminator returns true by default for unregistered operations
(presumably so that they don't break verification), and so they don't
get printed.

This change should make it so that these ops always get printed.