transformations: (lower-csl-wrapper) Add params_as_consts flag (#3324)
CSL module wrapper params are by default lowered to CSL params, which
can have a value specified in code that can be overridden manually by
providing command line inputs. If this is not desired, module wrapper
params can be lowered to constants instead by setting the
`params_as_consts` flag. For this to happen, the module wrapper params
must be numerical and have a specified default value.

Use this flag for programs if the generated program does not support
arbitrary param values.

---------

Co-authored-by: n-io <n-io@users.noreply.github.com>