API: use OpTraits to define traits on operations (#3363)
Using `@property` + `@classmethod` is deprecated since Python 3.11, and
removed in 3.13. The only place wehere we rely on this working is in the
traits mechanism, that does a bit of spooky aliasing of the traits class
variable to the OpDef traits property. This makes the following changes:
- changes the API of operation definitions to use OpTraits directly (in
line with how the other ClassVars are set on operations)
 - unifies the lazy/eager traits definitions (always just OpTraits)
- aliases the ClassVar and traits property so that if one is mutated the
other is also
 - adds a test for adding a trait to an operation