misc: raise error if variadic results types aren't referenced in assembly format (#3416)
The number of results is not passed in when parsing operations. In the
generic format, the type of the operation always specifies the types of
the results, and `resultSegmentSizes` specifies the ranges of of the
results if multiple are variadic. In order to support variadic results,
the types an length of all variadic results must be present in the
custom syntax.