core: new inference system for constraints (#3456)
Refactors inference for the irdl constraint system. 
The key addition is that each constraint now returns a
dictionary of "extractors" which retrieve the value of a
variable from the given attribute. This allows inference to be decoupled
entirely from verification, and inference should no longer be quite as
fragile/maybe slightly quicker.