dialects: (stream) simplify constr helper on stream attributes (#3473)
I'm working towards removing the stream dialect, and would like to
simplify it as much as possible before copying to both memref_stream and
snitch_stream. I don't think the constr overload is pulling its weight,
it's much simpler just to provide a shorthand for the BaseAttr
constraint, and only return a ParamAttrConstraint from constr.