core: Walk regions for rewritings instead of ops (#3506)
Currently, the pattern rewriter walker walks an operation, and tries to
apply patterns on all contained operations, including itself. This is
not what we would want, as rewriting a toplevel operation is dangerous,
if that operation does not have a parent.

This commit changes the walker to walk regions instead of operations, so
it is clear that the toplevel operation will not be modified. This is
similar to what MLIR currently does in applyPatternsAndFoldGreedily.