API: add `Op` suffix to all operations in xDSL (#3522)
I recently received feedback that one of the confusing aspects of xDSL
is that it's not consistent in terms of operation class naming, unlike
MLIR. This PR changes this. To make these changes, I added a runtime
check to Operation's post_init, asserted that it ended in "Op", and
changed the class names to make everything pass, then reverted that
change, so this PR converts all the classes that are initialized at some
point in our tests to this naming convention. We might want to add that
check in the future, but I wanted to discuss this in a separate PR.


This was the check I used:
``` python
        assert (
            "cmath" in self.name or self.__class__.__name__[-2:] == "Op"
        ), self.__class__.__name__
```