core: Fix extractor logic for optional and default-valued properties (#3525)
The extractor used in the assembly parser assumed that all
properties/attributes were present. We should instead only propagate the
extractors from the property constraint if the property is non-optional
or there is a default value. The default_value also needs to be given to
the extractor for when the property is not present.

Adds 2 tests to demonstrate this.